removed randset()

+++ aulib/melgen/randmel_gens.h
@@ -14,8 +14,10 @@ struct melody_temperley_params {
 	int nnts {100};
 	double p_major {0.88};
 	int sz_scdpool {120};  // scd's on:  [0 sz_scdpool)
-	int CP_mean {48};
-	int CP_stdev {13};
+	double CP_mean {48.0};
+	double CP_stdev {13.0};
+	//int CP_mean {48};
+	//int CP_stdev {13};
 	int RP_stdev {5};
 	int PP_stdev {7};   // double PP_stdev = 7.2;
 };  //TODO:  These stdev and mean params should be doubles...
